Rename existing tests to have SQLite suffix
These tests of the feature/session manager depend on SQLite.  A future test
class will exercise against SQL Server.
