+++ tests/Lussatite.FeatureManagement.Net6.Tests/SessionManagers/Sql/CachedSqlSessionManagerSQLiteTests.cs
@@ -1,20 +1,20 @@
 using System.Threading.Tasks;
 using LazyCache;
-using Lussatite.FeatureManagement.Net48.Tests.Testing.SQLite;
+using Lussatite.FeatureManagement.Net6.Tests.Testing.SQLite;
 using Lussatite.FeatureManagement.SessionManagers;
 using TestCommon.Standard;
 using TestCommon.Standard.SQLite;
 using Xunit;
 
-namespace Lussatite.FeatureManagement.Net48.Tests.SessionManagers.Sql
+namespace Lussatite.FeatureManagement.Net6.Tests.SessionManagers.Sql
 {
     [Collection(nameof(SQLiteDatabaseCollection))]
-    public class CachedSqlSessionManagerTests
+    public class CachedSqlSessionManagerSQLiteTests
     {
         private readonly SQLiteDatabaseFixture _dbFixture;
         private readonly IAppCache _appCache = new CachingService();
 
-        public CachedSqlSessionManagerTests(SQLiteDatabaseFixture dbFixture)
+        public CachedSqlSessionManagerSQLiteTests(SQLiteDatabaseFixture dbFixture)
         {
             _dbFixture = dbFixture;
         }
@@ -37,10 +37,11 @@ namespace Lussatite.FeatureManagement.Net48.Tests.SessionManagers.Sql
             Assert.Null(result);
         }
 
+
         [Theory]
-        [InlineData(null, "Net48_A125a_FeatureSetToNull", null)]
-        [InlineData(false, "Net48_A125b_FeatureSetToFalse", false)]
-        [InlineData(true, "Net48_A125c_FeatureSetToTrue", true)]
+        [InlineData(null, "Net6_A125a_FeatureSetToNull", null)]
+        [InlineData(false, "Net6_A125b_FeatureSetToFalse", false)]
+        [InlineData(true, "Net6_A125c_FeatureSetToTrue", true)]
         public async Task Return_expected_for_inserted_key_value(
             bool? expected,
             string featureName,
@@ -58,9 +59,9 @@ namespace Lussatite.FeatureManagement.Net48.Tests.SessionManagers.Sql
         }
 
         [Theory]
-        [InlineData(null, "Net48_A349x_FeatureSetToNull", null)]
-        [InlineData(false, "Net48_A349y_FeatureSetToFalse", false)]
-        [InlineData(true, "Net48_A349z_FeatureSetToTrue", true)]
+        [InlineData(null, "Net6_A349x_FeatureSetToNull", null)]
+        [InlineData(false, "Net6_A349y_FeatureSetToFalse", false)]
+        [InlineData(true, "Net6_A349z_FeatureSetToTrue", true)]
         public async Task Return_expected_for_SetNullableValue(
             bool? expected,
             string featureName,
@@ -78,9 +79,9 @@ namespace Lussatite.FeatureManagement.Net48.Tests.SessionManagers.Sql
         }
 
         [Theory]
-        [InlineData(null, "Net48_A359jx_FeatureSetToNull", null)]
-        [InlineData(false, "Net48_A359ky_FeatureSetToFalse", false)]
-        [InlineData(true, "Net48_A359lz_FeatureSetToTrue", true)]
+        [InlineData(null, "Net6_A359jx_FeatureSetToNull", null)]
+        [InlineData(false, "Net6_A359ky_FeatureSetToFalse", false)]
+        [InlineData(true, "Net6_A359lz_FeatureSetToTrue", true)]
         public async Task Return_expected_for_SetValue(
             bool? expected,
             string featureName,
@@ -102,7 +103,7 @@ namespace Lussatite.FeatureManagement.Net48.Tests.SessionManagers.Sql
         public async Task Exercise_SetNullableValue()
         {
             var sut = CreateSut();
-            const string baseName = "Net48_C997_ExerciseRepeatedly";
+            const string baseName = "Net6_C997_ExerciseRepeatedly";
             const int maxIterations = 1500;
             for (var i = 0; i < maxIterations; i++)
             {
@@ -119,7 +120,7 @@ namespace Lussatite.FeatureManagement.Net48.Tests.SessionManagers.Sql
         public async Task Exercise_SetValue()
         {
             var sut = CreateSut();
-            const string baseName = "Net48_C877_ExerciseRepeatedly";
+            const string baseName = "Net6_C877_ExerciseRepeatedly";
             const int maxIterations = 1500;
             for (var i = 0; i < maxIterations; i++)
             {
