+++ tests/Lussatite.FeatureManagement.NetCore31.Tests/SessionManagers/Sql/CachedSqlSessionManagerSQLiteTests.cs
@@ -1,27 +1,30 @@
 using System.Threading.Tasks;
-using Lussatite.FeatureManagement.Net48.Tests.Testing.SQLite;
+using LazyCache;
+using Lussatite.FeatureManagement.NetCore31.Tests.Testing.SQLite;
 using Lussatite.FeatureManagement.SessionManagers;
 using TestCommon.Standard;
 using TestCommon.Standard.SQLite;
 using Xunit;
 
-namespace Lussatite.FeatureManagement.Net48.Tests.SessionManagers.Sql
+namespace Lussatite.FeatureManagement.NetCore31.Tests.SessionManagers.Sql
 {
     [Collection(nameof(SQLiteDatabaseCollection))]
-    public class SqlSessionManagerTests
+    public class CachedSqlSessionManagerSQLiteTests
     {
         private readonly SQLiteDatabaseFixture _dbFixture;
+        private readonly IAppCache _appCache = new CachingService();
 
-        public SqlSessionManagerTests(SQLiteDatabaseFixture dbFixture)
+        public CachedSqlSessionManagerSQLiteTests(SQLiteDatabaseFixture dbFixture)
         {
             _dbFixture = dbFixture;
         }
 
-        private SqlSessionManager CreateSut()
+        private CachedSqlSessionManager CreateSut()
         {
             var settings = _dbFixture.SqlSessionManagerSettings;
 
-            return new SqlSessionManager(
+            return new CachedSqlSessionManager(
+                cache: _appCache,
                 settings: settings
                 );
         }
@@ -34,10 +37,11 @@ namespace Lussatite.FeatureManagement.Net48.Tests.SessionManagers.Sql
             Assert.Null(result);
         }
 
+
         [Theory]
-        [InlineData(null, "Net48_A125a_FeatureSetToNull", null)]
-        [InlineData(false, "Net48_A125b_FeatureSetToFalse", false)]
-        [InlineData(true, "Net48_A125c_FeatureSetToTrue", true)]
+        [InlineData(null, "NetCore31_A125a_FeatureSetToNull", null)]
+        [InlineData(false, "NetCore31_A125b_FeatureSetToFalse", false)]
+        [InlineData(true, "NetCore31_A125c_FeatureSetToTrue", true)]
         public async Task Return_expected_for_inserted_key_value(
             bool? expected,
             string featureName,
@@ -55,9 +59,9 @@ namespace Lussatite.FeatureManagement.Net48.Tests.SessionManagers.Sql
         }
 
         [Theory]
-        [InlineData(null, "Net48_A129x_FeatureSetToNull", null)]
-        [InlineData(false, "Net48_A129y_FeatureSetToFalse", false)]
-        [InlineData(true, "Net48_A129z_FeatureSetToTrue", true)]
+        [InlineData(null, "NetCore31_A349x_FeatureSetToNull", null)]
+        [InlineData(false, "NetCore31_A349y_FeatureSetToFalse", false)]
+        [InlineData(true, "NetCore31_A349z_FeatureSetToTrue", true)]
         public async Task Return_expected_for_SetNullableValue(
             bool? expected,
             string featureName,
@@ -75,9 +79,9 @@ namespace Lussatite.FeatureManagement.Net48.Tests.SessionManagers.Sql
         }
 
         [Theory]
-        [InlineData(null, "Net48_A139jx_FeatureSetToNull", null)]
-        [InlineData(false, "Net48_A139ky_FeatureSetToFalse", false)]
-        [InlineData(true, "Net48_A139lz_FeatureSetToTrue", true)]
+        [InlineData(null, "NetCore31_A359jx_FeatureSetToNull", null)]
+        [InlineData(false, "NetCore31_A359ky_FeatureSetToFalse", false)]
+        [InlineData(true, "NetCore31_A359lz_FeatureSetToTrue", true)]
         public async Task Return_expected_for_SetValue(
             bool? expected,
             string featureName,
@@ -99,8 +103,8 @@ namespace Lussatite.FeatureManagement.Net48.Tests.SessionManagers.Sql
         public async Task Exercise_SetNullableValue()
         {
             var sut = CreateSut();
-            const string baseName = "Net48_A997_ExerciseRepeatedly";
-            const int maxIterations = 500;
+            const string baseName = "NetCore31_C997_ExerciseRepeatedly";
+            const int maxIterations = 1500;
             for (var i = 0; i < maxIterations; i++)
             {
                 var callSet = Rng.GetInteger(0, 20) == 0;
@@ -116,8 +120,8 @@ namespace Lussatite.FeatureManagement.Net48.Tests.SessionManagers.Sql
         public async Task Exercise_SetValue()
         {
             var sut = CreateSut();
-            const string baseName = "Net48_A877_ExerciseRepeatedly";
-            const int maxIterations = 500;
+            const string baseName = "NetCore31_C877_ExerciseRepeatedly";
+            const int maxIterations = 1500;
             for (var i = 0; i < maxIterations; i++)
             {
                 var callSet = Rng.GetInteger(0, 20) == 0;
